requirements.txt
azure-identity
azure-mgmt-compute